+++ app/controllers/shop/api/orders_controller.rb
@@ -169,7 +169,10 @@ module Shop
       end
 
       def order_params
-        params.permit(:name, :email, :phone, :comment, :is_car_pickup, :car_number, :promo_code, :payment_method, :pickup_time)
+        params.permit(
+          :name, :email, :phone, :comment, :is_car_pickup, :car_number,
+          :promo_code, :payment_method, :pickup_time, :client_order_uuid
+        )
       end
 
       def try_reconnect_from_params!